Rename view path -> pages & move some components into components path
