+++ lib/components/custom_scaffold_builder.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'components/custom_app_bar.dart';
+import 'custom_app_bar.dart';
 
 Widget customScaffoldBuilder(BuildContext context, Widget title, Widget child) {
   return Scaffold(
